--- pages/developers.js
@@ -1,11 +1,0 @@
-import ScrapdLayout from '../components/scrapd-layout';
-import ScrapdApdViewContainer from '../components/scrapd-developers-view/container';
-import { connect } from 'react-redux';
-
-const Developers = () => (
-  <ScrapdLayout>
-    <ScrapdApdViewContainer />
-  </ScrapdLayout>
-);
-
-export default connect()(Developers);
